LPS-99657 Class that implements StrutsAction should be in package *.struts
